move setting extension to relative class and sort usings
