+++ src/blueprints/Do.Blueprints.Service.Application/MockOverrider/MockOverriderExtensions.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Do;
 
-public static partial class MockOverriderExtensions
+public static class MockOverriderExtensions
 {
     public static void AddMockOverrider(this List<IFeature> source, Func<MockOverriderConfigurator, IFeature<MockOverriderConfigurator>> configure) => source.Add(configure(new()));
